+++ test/helper.rb
@@ -15,6 +15,7 @@ rescue LoadError
 end
 
 require 'minitest/autorun'
+require 'wrapture'
 
 def get_include_list(filename)
   includes = []
@@ -29,7 +30,7 @@ end
 
 def validate_declaration_file(spec)
   filename = "#{spec['name']}.hpp"
-  class_includes = spec['includes'] || []
+  class_includes = Wrapture::ClassSpec.normalize_spec_hash(spec)['includes']
 
   includes = get_include_list filename
 
@@ -42,7 +43,7 @@ end
 
 def validate_definition_file(spec)
   filename = "#{spec['name']}.cpp"
-  class_includes = spec['includes'] || []
+  class_includes = Wrapture::ClassSpec.normalize_spec_hash(spec)['includes']
 
   includes = get_include_list filename