+++ ChangeLog.md
@@ -9,10 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Python class generation.
  - Perl class generation.
 
-## [0.3.0] - 2019-07-07
+## [0.3.0] - 2019-07-08
 ### Added
  - Examples of basic usage and features.
  - Option to give a class-level include list.
+ - Include lists may now be either a single string or a list.
 
 ### Fixed
  - `bin/wrapture` is now executable.